requirements.txt
# Core ML/DL Frameworks
torch>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Reinforcement Learning
stable-baselines3>=2.1.0
gymnasium>=0.29.0
ray[rllib]>=2.7.0

# Deep Learning Utilities
transformers>=4.30.0
einops>=0.7.0

# Data Processing
pandas>=2.0.0
polars>=0.19.0
numpy>=1.24.0
scipy>=1.11.0
pyarrow>=13.0.0

# Feature Engineering
category-encoders>=2.6.0
feature-engine>=1.6.0

# Optimization
pymoo>=0.6.0
cvxpy>=1.4.0
pulp>=2.7.0
ortools>=9.7.0

# Hyperparameter Tuning
optuna>=3.3.0
ray[tune]>=2.7.0

# API Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.3.0
python-multipart>=0.0.6

# Database & Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pymongo>=4.5.0

# Message Queue & Streaming
kafka-python>=2.0.2
celery>=5.3.0
pika>=1.3.0

# Monitoring & Logging
mlflow>=2.7.0
wandb>=0.15.0
prometheus-client>=0.17.0
structlog>=23.1.0
python-json-logger>=2.0.7

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
hypothesis>=6.87.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.0
isort>=5.12.0
pre-commit>=3.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
click>=8.1.0
joblib>=1.3.0
cloudpickle>=2.2.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.4
sktime>=0.24.0

# Network Analysis
networkx>=3.1
igraph>=0.10.0

# Profiling & Performance
line-profiler>=4.1.0
memory-profiler>=0.61.0
py-spy>=0.3.14

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0
pdoc>=14.1.0

# Security
cryptography>=41.0.0
